+++ build.gradle
@@ -23,29 +23,30 @@ repositories {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-    implementation 'org.apache.httpcomponents:httpclient:4.5.13'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-mail'
     implementation 'org.springframework.boot:spring-boot-starter-security'
     compileOnly 'org.projectlombok:lombok'
     runtimeOnly 'mysql:mysql-connector-java'
-    annotationProcessor 'org.projectlombok:lombok'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.security:spring-security-test'
     testImplementation 'org.junit.jupiter:junit-jupiter-engine'
+    implementation 'org.apache.httpcomponents:httpclient:4.5.13'
     implementation 'mysql:mysql-connector-java:8.0.28'
     implementation 'org.webjars:jquery:3.4.1'
     implementation 'org.webjars:bootstrap:4.3.1'
     implementation 'org.webjars:webjars-locator-core'
     implementation 'org.thymeleaf:thymeleaf:3.1.2.RELEASE'
-    implementation("nz.net.ultraq.thymeleaf:thymeleaf-layout-dialect")
-    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
-
+    implementation 'nz.net.ultraq.thymeleaf:thymeleaf-layout-dialect'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:2.5.5'
+    implementation 'org.hibernate:hibernate-core:6.4.1.Final'
+    implementation 'javax.xml.bind:jaxb-api:2.3.1'
+    implementation 'org.javassist:javassist:3.25.0-GA'
 }
+
 test {
     useJUnitPlatform()
 }
