--- src/main/java/org/example/AppConfig.java
@@ -1,21 +1,0 @@
-package org.example;
-
-import org.example.back.remote.LibraryService.Remote.GutendexService;
-import org.example.back.remote.LibraryService.Remote.LibraryServiceRemoteDataSource;
-import org.example.back.remote.LibraryService.Remote.LibraryServiceRemoteDataSourceImp;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class AppConfig {
-
-    @Bean
-    public LibraryServiceRemoteDataSource libraryServiceRemoteDataSource() {
-        return new LibraryServiceRemoteDataSourceImp();
-    }
-
-    @Bean
-    public GutendexService gutendexService(LibraryServiceRemoteDataSource remoteDataSource) {
-        return new GutendexService(remoteDataSource);
-    }
-}
