--- src/main/java/org/example/AppController.java
@@ -1,58 +1,0 @@
-package org.example;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-
-import java.util.List;
-
-@Controller
-public class AppController {
-
-    @Autowired
-    private UserRepository userRepo;
-
-    @GetMapping("")
-    public String viewHomePage() {
-        return "index";
-    }
-    @GetMapping("/login")
-    public String showLoginPage() {
-        return "login";
-    }
-    @PostMapping("/login")
-    public String handleLoginPost() {
-        return "redirect:/dashboard";
-    }
-    @GetMapping("/register")
-    public String showRegistrationForm(Model model) {
-        model.addAttribute("user", new User());
-
-        return "signup_form";
-    }
-    @GetMapping("/dashboard")
-    public String showDashboard() {
-        return "dashboard";
-    }
-    @PostMapping("/process_register")
-    public String processRegister(User user) {
-        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-        String encodedPassword = passwordEncoder.encode(user.getPassword());
-        user.setPassword(encodedPassword);
-
-        userRepo.save(user);
-
-        return "register_success";
-    }
-
-    @GetMapping("/users")
-    public String listUsers(Model model) {
-        List<User> listUsers = userRepo.findAll();
-        model.addAttribute("listUsers", listUsers);
-
-        return "users";
-    }
-}
