--- src/main/java/org/example/Main.java
@@ -1,16 +1,0 @@
-package org.example;
-
-import org.example.back.remote.LibraryService.Remote.GutendexService;
-import org.example.back.remote.LibraryService.Remote.LibraryServiceRemoteDataSource;
-import org.example.back.remote.LibraryService.Remote.LibraryServiceRemoteDataSourceImp;
-
-public class Main {
-    public static void main(String[] args) {
-        LibraryServiceRemoteDataSource remoteDataSource = new LibraryServiceRemoteDataSourceImp();
-        GutendexService gutendexService = new GutendexService(remoteDataSource);
-
-        String queryParameters = "";
-
-        gutendexService.searchBooks(queryParameters);
-    }
-}
