--- src/main/java/org/example/UserController.java
@@ -1,19 +1,0 @@
-package org.example;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/users")
-public class UserController {
-    @Autowired
-    private UserRepository userRepository;
-
-    @PostMapping("/register")
-    public User registerUser(@RequestBody User user) {
-        return userRepository.save(user);
-    }
-}
