--- src/main/java/org/example/back/remote/LibraryService/Remote/LibraryServiceRemoteDataSource.java
@@ -1,7 +1,0 @@
-package org.example.back.remote.LibraryService.Remote;
-
-import org.example.back.remote.LibraryService.Remote.Model.BookList;
-
-public interface LibraryServiceRemoteDataSource {
-    BookList getBookData(String queryParameters);
-}
