--- src/main/java/org/example/back/remote/LibraryService/Remote/Model/Book.java
@@ -1,105 +1,0 @@
-package org.example.back.remote.LibraryService.Remote.Model;
-
-import java.util.List;
-
-public class Book {
-    public int id;
-    public String title;
-    public List<String> subjects;
-    public List<Person> authors;
-    public List<Person> translators;
-    public List<String> bookshelves;
-    public List<String> languages;
-    public Boolean copyright;
-    public String media_type;
-    public Format formats;
-    public int download_count;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public List<String> getSubjects() {
-        return subjects;
-    }
-
-    public void setSubjects(List<String> subjects) {
-        this.subjects = subjects;
-    }
-
-    public List<Person> getAuthors() {
-        return authors;
-    }
-
-    public void setAuthors(List<Person> authors) {
-        this.authors = authors;
-    }
-
-    public List<Person> getTranslators() {
-        return translators;
-    }
-
-    public void setTranslators(List<Person> translators) {
-        this.translators = translators;
-    }
-
-    public List<String> getBookshelves() {
-        return bookshelves;
-    }
-
-    public void setBookshelves(List<String> bookshelves) {
-        this.bookshelves = bookshelves;
-    }
-
-    public List<String> getLanguages() {
-        return languages;
-    }
-
-    public void setLanguages(List<String> languages) {
-        this.languages = languages;
-    }
-
-    public Boolean getCopyright() {
-        return copyright;
-    }
-
-    public void setCopyright(Boolean copyright) {
-        this.copyright = copyright;
-    }
-
-    public String getMedia_type() {
-        return media_type;
-    }
-
-    public void setMedia_type(String media_type) {
-        this.media_type = media_type;
-    }
-
-    public Format getFormats() {
-        return formats;
-    }
-
-    public void setFormats(Format formats) {
-        this.formats = formats;
-    }
-
-    public int getDownload_count() {
-        return download_count;
-    }
-
-    public void setDownload_count(int download_count) {
-        this.download_count = download_count;
-    }
-}
