--- src/main/java/org/example/back/remote/LibraryService/Remote/Model/ServiceError.java
@@ -1,13 +1,0 @@
-package org.example.back.remote.LibraryService.Remote.Model;
-
-public class ServiceError {
-    public String detail;
-
-    public String getDetail() {
-        return detail;
-    }
-
-    public void setDetail(String detail) {
-        this.detail = detail;
-    }
-}
