+++ src/main/java/org/example/controller/UserController.java
@@ -1,0 +1,21 @@
+package org.example.controller;
+
+import org.example.entity.User;
+import org.example.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/users")
+public class UserController {
+    @Autowired
+    private UserRepository userRepository;
+
+    @PostMapping("/register")
+    public User registerUser(@RequestBody User user) {
+        return userRepository.save(user);
+    }
+}
