+++ src/main/java/org/example/repository/RoleRepository.java
@@ -1,0 +1,9 @@
+package org.example.repository;
+
+import org.example.entity.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoleRepository extends JpaRepository<Role, Long> {
+
+    Role findByName(String name);
+}
