+++ src/main/java/org/example/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package org.example.repository;
+
+import org.example.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    User findByEmail(String email);
+
+}
