+++ src/main/java/org/example/service/UserService.java
@@ -1,0 +1,14 @@
+package org.example.service;
+
+import org.example.dto.UserDto;
+import org.example.entity.User;
+
+import java.util.List;
+
+public interface UserService {
+    void saveUser(UserDto userDto);
+
+    User findUserByEmail(String email);
+
+    List<UserDto> findAllUsers();
+}
