+++ src/main/java/org/example/service/searchApi/remote/LibraryService/Remote/LibraryServiceRemoteDataSource.java
@@ -1,0 +1,7 @@
+package org.example.service.searchApi.remote.LibraryService.Remote;
+
+import org.example.service.searchApi.remote.LibraryService.Remote.Model.BookList;
+
+public interface LibraryServiceRemoteDataSource {
+    BookList getBookData(String queryParameters);
+}
