+++ src/main/java/org/example/service/searchApi/remote/LibraryService/Remote/Model/BookList.java
@@ -1,0 +1,42 @@
+package org.example.service.searchApi.remote.LibraryService.Remote.Model;
+
+import java.util.List;
+
+public class BookList {
+    public int count;
+    public String next;
+    public String previous;
+    public List<Book> results;
+
+    public int getCount() {
+        return count;
+    }
+
+    public void setCount(int count) {
+        this.count = count;
+    }
+
+    public String getNext() {
+        return next;
+    }
+
+    public void setNext(String next) {
+        this.next = next;
+    }
+
+    public String getPrevious() {
+        return previous;
+    }
+
+    public void setPrevious(String previous) {
+        this.previous = previous;
+    }
+
+    public List<Book> getResults() {
+        return results;
+    }
+
+    public void setResults(List<Book> results) {
+        this.results = results;
+    }
+}
