+++ src/main/java/org/example/service/searchApi/remote/LibraryService/Remote/Model/Person.java
@@ -1,0 +1,31 @@
+package org.example.service.searchApi.remote.LibraryService.Remote.Model;
+
+public class Person {
+    public Integer birth_year;
+    public Integer death_year;
+    public String name;
+
+    public Integer getBirth_year() {
+        return birth_year;
+    }
+
+    public void setBirth_year(Integer birth_year) {
+        this.birth_year = birth_year;
+    }
+
+    public Integer getDeath_year() {
+        return death_year;
+    }
+
+    public void setDeath_year(Integer death_year) {
+        this.death_year = death_year;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
