+++ src/main/java/org/example/service/searchApi/remote/LibraryService/Remote/Model/ServiceError.java
@@ -1,0 +1,13 @@
+package org.example.service.searchApi.remote.LibraryService.Remote.Model;
+
+public class ServiceError {
+    public String detail;
+
+    public String getDetail() {
+        return detail;
+    }
+
+    public void setDetail(String detail) {
+        this.detail = detail;
+    }
+}
